edge-cloud-coordinator/requirements.txt
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0
websockets>=10.0
fastapi>=0.85.0
uvicorn>=0.18.0
pydantic>=2.0.0
kafka-python>=2.0.2
pika>=1.3.0
requests>=2.27.0

# 熔断器
pybreaker>=1.0.0

# JWT认证（标准库）
pyjwt[crypto]>=2.8.0

# 加密
cryptography>=41.0.0
